+++ Eduplatform_site/account/models.py
@@ -2,7 +2,7 @@
 from django.db import models
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 from django.utils.translation import gettext_lazy as _
-from .mixins import DateTimeMixinModel
+from Eduplatform_site.mixins import DateTimeMixinModel
 from .managers import CustomUserManager
